Module Sofo.Convenience

Contains convenience functions.

val print : Sexplib0.Sexp.t -> Base.unit

val trans_2d : Torch.Tensor.t -> Torch.Tensor.t

Transpose a along the first two dimensions.

val extend_one : Forward_torch.Maths.t -> Forward_torch.Maths.t

Extend the shape of a from shape to shape; 1.

val first_dim : Torch.Tensor.t -> Base.int

Get first dimension of a.

val a_b_trans : Torch.Tensor.t -> Torch.Tensor.t -> Torch.Tensor.t

a *@ (transpose b)

val a_trans_b : Torch.Tensor.t -> Torch.Tensor.t -> Torch.Tensor.t

(transpose a) *@ b

val all_dims_but_first : Torch.Tensor.t -> int Base.List.t

Returns an int list starting from 1 and ending at the last dim of a.

val sum_except_dim0 : Torch.Tensor.t -> Torch.Tensor.t

Sum all elements in x except for dimension 0.

val top_eigval : Torch.Tensor.t -> float

Get largest eigenvalue of x

val gaussian_tensor_2d_normed : device:Torch_core.Device.t -> kind:Torch_core.Kind.packed -> a:int -> b:int -> sigma:Base.Float.t -> Torch.Tensor.t

Return a Random gaussian tensor of standard deviation sigma of kind normalised on rows on device with shape a;b.

val num_train_loops : full_batch_size:Base.Int.t -> batch_size:Base.Int.t -> Base.Int.t -> Base.Int.t

Calculate num of training loops given full_batch_size, batch_size and num_epochs_to_run.

val epoch_of : full_batch_size:int -> batch_size:int -> int -> Base.Float.t

Calculate current epoch number given full_batch_size, batch_size and iteration number.

val calc_out_height_width_channel : padding:(Base.Int.t * Base.Int.t) -> dilation:(Base.Int.t * Base.Int.t) -> in_out_channels_kernel:('a * 'b * Base.Int.t * Base.Int.t) -> stride:(Base.Int.t * Base.Int.t) -> in_height:Base.Int.t -> in_width:Base.Int.t -> Base.Int.t * Base.Int.t * 'b

Calculate output height, width and channel after a conv2d operation.