Module Sofo.Loss

Loss function to be minimised.

Loss modules.

module type T = sig ... end

Module for loss function.

module type Reward = sig ... end

Module for reward function, where loss is usually defined as negative reward.

module MSE (_ : sig ... end) : T with type labels = Torch.Tensor.t and type 'a with_args = 'a

Mean squared loss.

module CE (_ : sig ... end) : T with type labels = Torch.Tensor.t and type 'a with_args = 'a

Cross-entropy loss.

module Weighted_MSE (_ : sig ... end) : T with type labels = Torch.Tensor.t and type 'a with_args = weights:Torch.Tensor.t -> 'a

Weighted mean squared loss, with the loss and sketched GGN multiplied by the scaling_factor.

module RL_loss : Reward with type 'a with_args = 'a

Negative of utility/reward.