Module Optimizer.Adam

Adam optimizer

Parameters

module W : Wrapper.T

Signature

module W = W

W is the wrapper type that defines the forward computational graph.

type ('a, 'b) config = ('a, 'b) Config.Adam.t

type state

type ('c, _, _) init_opts = W.P.tagged -> 'c

val params : state -> W.P.tagged

Extract parameters from state.

val init : (state, 'a, 'b) init_opts

Initialise state from parameters.

val step : config:('a, 'b) config -> state:state -> data:W.data -> args:W.args -> float * state

Given config, current state, data and additional args, return loss and updated state.