Module T.W

W is the wrapper type that defines the forward computational graph.

module P : Forward_torch.Prms.T

type data

Type of data to be passed into the network.

type args

Type of any additional arguments.

val f : update: [ `loss_only of 'a -> Forward_torch.Maths.t option -> 'a | `loss_and_ggn of 'a -> (Forward_torch.Maths.t * Torch.Tensor.t option) option -> 'a ] -> data:data -> init:'a -> args:args -> P.M.t -> 'a

Given an update function update, data, the initial value of the losses and ggn init and args, return the final losses and ggn.