Module Sofo.Wrapper

Wrappers define the forward computational graph.

Two common wrappers.

module type Function = sig ... end

Forward function.

module type Recurrent_function = sig ... end

Recurrent function.

module type T = sig ... end

Basic wrapper type, which is the forward computational graph of the network.

module Feedforward (F : Function) (L : Loss.T with type 'a with_args = 'a) : T with module P = F.P and type data = F.input * L.labels and type args = unit

Feedforward network where F is the forward function and L is the loss module.

module Recurrent (F : Recurrent_function) (L : Loss.T with type 'a with_args = 'a) : T with module P = F.P and type data = (F.input * L.labels option) list and type args = Torch.Tensor.t

Recurrent network where F is the one-step recurrent function and L is the loss module.