Module Wrapper.Recurrent

Recurrent network where F is the one-step recurrent function and L is the loss module.

Parameters

module F : Recurrent_function

module L : Loss.T with type 'a with_args = 'a

Signature

module P = F.P

type data = (F.input * L.labels option) list

Type of data to be passed into the network.

type args = Torch.Tensor.t

Type of any additional arguments.

val f : update: [ `loss_only of 'a -> Forward_torch.Maths.t option -> 'a | `loss_and_ggn of 'a -> (Forward_torch.Maths.t * Torch.Tensor.t option) option -> 'a ] -> data:data -> init:'a -> args:args -> P.M.t -> 'a

Given an update function update, data, the initial value of the losses and ggn init and args, return the final losses and ggn.